requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
web3==7.6.0
pydantic-core==2.33.1
pyrate-limiter==3.7.0
pydantic>=2.0,<3.0
pydantic-settings==2.9.1
pyyaml==6.0.2
structlog==25.3.0